ml-brain/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
scikit-learn==1.5.0
pandas==2.2.2
psycopg2-binary==2.9.9
joblib==1.4.2
numpy==1.26.4
python-dotenv==1.0.1
apscheduler==3.10.4
